backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
Werkzeug==3.0.1
firebase-admin==6.5.0
gunicorn==22.0.0
simple-websocket==1.1.0
